+++ l10n/de_DE.js
@@ -74,6 +74,9 @@ OC.L10N.register(
         'Overview': 'Überblick',
         'Drilldown': 'Aufriss',
         'Datasource simulation': 'Datasource simulieren',
-        'Link copied': 'Link kopiert'
+        'Link copied': 'Link kopiert',
+        'Bar': 'Balken',
+        'Text variables': 'Text Variablen',
+        'Text variables can be used in title and subheader.<br>They are replaced when the report is executed.': 'Text Variablen können im Berichtsnamen oder der Beschreibung verwendet werden.<br>Sie werden ersetzt, wenn der Bericht ausgeführt wird.'
     },
     '');